+++ src/types/electron.d.ts
@@ -1,0 +1,30 @@
+export interface UpdateInfo {
+  version: string;
+  releaseNotes?: string;
+}
+
+export interface DownloadProgress {
+  percent: number;
+  bytesPerSecond: number;
+  transferred: number;
+  total: number;
+}
+
+export interface ElectronAPI {
+  openExternal: (url: string) => void;
+  checkForUpdates: () => void;
+  downloadUpdate: () => void;
+  installUpdate: () => void;
+  onUpdateAvailable: (callback: (info: UpdateInfo) => void) => void;
+  onUpdateNotAvailable: (callback: (info: UpdateInfo) => void) => void;
+  onUpdateDownloaded: (callback: (info: UpdateInfo) => void) => void;
+  onDownloadProgress: (callback: (progress: DownloadProgress) => void) => void;
+  onUpdateError: (callback: (error: string) => void) => void;
+  removeUpdateListeners: () => void;
+}
+
+declare global {
+  interface Window {
+    electronAPI?: ElectronAPI;
+  }
+}
